+++ src/components/home.tsx
@@ -2,7 +2,7 @@ import { useState, useEffect } from "react";
 import styles from "../../styles/home.module.css";
 import { useRouter } from "next/router";
 import { Popup } from "./popup";
-import { isContributor } from "../../utils/Scraper";
+import { useSession } from "next-auth/react";
 
 interface trackInterface {
   title: string;
@@ -26,6 +26,8 @@ export default function Home() {
     id: "null",
   });
   const [isLoading, setIsLoading] = useState<boolean>(true);
+  const { status } = useSession();
+  
   function getRepoName(url: string) {
     const ownerRe = new RegExp("(?<=https://github.com/).+?(?=/)", "g");
     const owner = ownerRe.exec(url);
@@ -75,6 +77,10 @@ export default function Home() {
   };
 
   async function deleteTrack(id: string) {
+    if (status === "unauthenticated") {
+      alert("Failed to delete!\nPlease Sign in");
+      return;
+    }
     const response = await fetch(`/api/tracks/${id}`, {
       method: "DELETE",
       headers: {
@@ -83,8 +89,16 @@ export default function Home() {
       },
     });
     const data = await response.json();
+    console.log("Status code is: ", response.status);
     if (!data.success) {
-      alert("Failed to delete");
+      const statusCode = response.status;
+      if (statusCode === 401) {
+        alert("Failed to delete!\nPlease Sign in");
+      } else if (statusCode === 403) {
+        alert("Failed to delete!\nYou don't have permission to delete");
+      } else if (statusCode === 403) {
+        alert("Failed to delete!\nBad request");
+      }
     } else {
       alert("Deleted successfully!");
     }
